Refactor match history code into matchHistory.js
